20.08.02 refactor auth, s3
