+++ Chinstagram/src/main/java/com/landvibe/chinstagram/config/WebConfig.java
@@ -1,4 +1,4 @@
-package com.landvibe.chinstagram;
+package com.landvibe.chinstagram.config;
 
 import com.landvibe.chinstagram.jwt.JwtAuthInterceptor;
 import org.springframework.beans.factory.annotation.Autowired;
