+++ Chinstagram/src/main/java/com/landvibe/chinstagram/models/Content.java
@@ -46,5 +46,4 @@ public class Content {
         this.createTime = createTime;
         this.updateTime = updateTime;
     }
-
 }
