+++ Chinstagram/src/main/java/com/landvibe/chinstagram/models/Image.java
@@ -19,5 +19,4 @@ public class Image {
 
     @Column
     private String path;
-
 }
