+++ Chinstagram/src/main/java/com/landvibe/chinstagram/models/Profile.java
@@ -44,5 +44,4 @@ public class Profile {
     protected void onUpdate() {
         updateTime = LocalDateTime.now();
     }
-
 }
