+++ Chinstagram/src/main/java/com/landvibe/chinstagram/models/User.java
@@ -29,5 +29,4 @@ public class User {
     @OneToOne
     @JoinColumn(name = "profile_id")
     private Profile profile;
-
 }
