+++ Chinstagram/src/main/java/com/landvibe/chinstagram/services/UserService.java
@@ -6,7 +6,6 @@ import com.landvibe.chinstagram.repositories.ImageRepository;
 import com.landvibe.chinstagram.repositories.UserRepository;
 import com.landvibe.chinstagram.uploader.S3Uploader;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.cache.CacheManager;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.HttpServerErrorException;
@@ -23,9 +22,6 @@ public class UserService {
 
     @Autowired
     private S3Uploader s3Uploader;
-
-    @Autowired
-    private CacheManager cacheManager;
 
     @Autowired
     private ImageRepository imageRepository;
