+++ DevelopLog/20.08.02.md
@@ -1,0 +1,12 @@
+# 20.08.02
+
+### 오늘 한 것
+
+- auth 인증 부분 완료
+  - 8/1일에 스터디하면서 코멘트 받은 부분들 수정
+- s3 업로드 부분 완료
+  - postman 이용해서 이미지 업로드 되는 것 확인 완료
+- postman으로 모든 응답 확인 완료
+
+### 더 해야할 것
+
